refactor: change `metricsql` funcs structure by folders
