+++ src/querybuilder/metricsql-functions/time/index.ts
@@ -1,0 +1,11 @@
+import { QueryBuilderOperationDef } from "../../shared/types";
+
+import { getSimpleTimeFunctions } from "./time";
+import { getTimezoneOffset } from "./timezoneOffset";
+
+export function getTimeFunctions(): QueryBuilderOperationDef[] {
+  return [
+    ...getSimpleTimeFunctions(),
+    getTimezoneOffset(),
+  ]
+}
